Commit 1
Rectified the issue with ifft in nidfft after 5 long hours.
It was a really juvenile mistake - instead of (fftw_complex) I
had typed (fftw_malloc) and it was throwing all sorts of errors
back at me. It was still fun :)

So now I have a header file with functions to read a particular column
of a file and to read 2 particular columns of a file.
Other than this I have a binary "nidfft" that can do an FFT of a 1D
data (with good user-friendly options) and output either the power or
the complex form of the transform. The same code can also be used to
invert a 2 column complex fft of real data from frequency domain to the
time domain.

Next stop, Filters and STFT.